% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_psem.R
\name{AIC.psem}
\alias{AIC.psem}
\title{Generalized function for SEM AIC(c) score}
\usage{
\method{AIC}{psem}(object, ..., aicc = FALSE)
}
\arguments{
\item{object}{a psem object}

\item{...}{additional arguments to AIC}

\item{aicc}{whether correction for small sample size should be applied. Default is \code{FALSE}}
}
\description{
Generalized function for SEM AIC(c) score
}
