% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all.vars_notrans}
\alias{all.vars_notrans}
\title{Get vector of untransformed variables}
\usage{
\method{all}{vars_notrans}(formula.)
}
\description{
Get vector of untransformed variables
}
\keyword{internal}
