% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all.vars_trans}
\alias{all.vars_trans}
\title{Get vector of transformed variables}
\usage{
\method{all}{vars_trans}(formula., smoothed = FALSE)
}
\description{
Get vector of transformed variables
}
\keyword{internal}
