% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.psem.R
\name{anova.psem}
\alias{anova.psem}
\title{ANOVA and chi-squared difference test for model comparison}
\usage{
\method{anova}{psem}(object, ...)
}
\arguments{
\item{object}{a \code{psem} object}

\item{...}{additional \code{psem} objects}

\item{test.type}{what kind of ANOVA is conducted. Default is type III}
}
\value{
an ANOVA table for a single model, a list of comparisons between multiple
models
}
\description{
Compute analysis of variance table for one or more structural equation models.
}
\examples{
mod1 <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley
)

# get type III Anova
anova(mod1)

# conduct LRT
mod2 <- psem(
  lm(rich ~ cover, data = keeley),
  lm(cover ~ firesev, data = keeley),
  age ~ 1,
  data = keeley
)

anova(mod1, mod2)

}
\seealso{
The model fitting function \code{\link{psem}}
}
\author{
Jon Lefcheck <LefcheckJ@si.edu>, Jarrett Byrnes <jarrett.byrnes@umb.edu>
}
