% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{as.psem}
\alias{as.psem}
\title{Convert list to psem object}
\usage{
as.psem(object, Class = "psem")
}
\arguments{
\item{object}{any \code{R} object}

\item{Class}{the name of the class to which \code{object} should be coerced}
}
\description{
Convert list to psem object
}
