% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{basisSet}
\alias{basisSet}
\title{Derivation of the basis set}
\usage{
basisSet(modelList, direction = NULL)
}
\arguments{
\item{modelList}{A list of structural equations.}

\item{direction}{A \code{vector} of claims defining the specific
directionality of independence claims; for use in special cases (see
Details).}
}
\value{
A \code{list} of independence claims.
}
\description{
Derivation of the set of independence claims--the basis set--for use in
evaluating the goodness-of-fit for piecewise structural equation models.
}
\examples{




}
\author{
Jon Lefcheck <jlefcheck@bigelow.org>
}
\references{
Shipley, Bill. "A new inferential test for path models based on
directed acyclic graphs." Structural Equation Modeling 7.2 (2000): 206-218.
}

