% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_psem.R
\name{BIC.psem}
\alias{BIC.psem}
\title{Generalized function for SEM BIC score}
\usage{
\method{BIC}{psem}(object, ...)
}
\arguments{
\item{object}{a psem object}

\item{...}{additional arguments to BIC}
}
\description{
Generalized function for SEM BIC score
}
