% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{checkData}
\alias{checkData}
\title{Check to see whether supplied data.frame matches model-extracted data}
\usage{
checkData(x)
}
\description{
Check to see whether supplied data.frame matches model-extracted data
}
\keyword{internal}
