% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{checkTransformations}
\alias{checkTransformations}
\title{Check to see whether variables exist as transformed and untransformed}
\usage{
checkTransformations(x)
}
\description{
Check to see whether variables exist as transformed and untransformed
}
\keyword{internal}
