% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dag.R
\name{Dag}
\alias{Dag}
\title{Generate adjacency matrix from list of structural equations}
\usage{
Dag(modelList)
}
\arguments{
\item{modelList}{A list of structural equations}
}
\description{
Generate adjacency matrix from list of structural equations
}
