% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{deparseInt}
\alias{deparseInt}
\title{Determines if we need to use emmeans or emtrends}
\usage{
deparseInt(coefName, model, catVars, vars)
}
\description{
Determines if we need to use emmeans or emtrends
}
\keyword{internal}
