% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldfunctions.R
\name{endogenous.reverse}
\alias{endogenous.reverse}
\title{Functions from version 1.2 of piecewiseSEM
For use in vignette()}
\usage{
endogenous.reverse(basis.set, modelList, add.vars = NULL)
}
\description{
Functions from version 1.2 of piecewiseSEM
For use in vignette()
}
