% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{filterExogenous}
\alias{filterExogenous}
\title{Filter relationships among exogenous variables from the basis set (ignoring add.vars)}
\usage{
filterExogenous(b, modelList, amat)
}
\description{
Filter relationships among exogenous variables from the basis set (ignoring add.vars)
}
\keyword{internal}
