% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{filterInteractions}
\alias{filterInteractions}
\title{Filter interactions from the d-sep tests}
\usage{
filterInteractions(b, interactions = FALSE)
}
\description{
Filter interactions from the d-sep tests
}
\keyword{internal}
