% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{filterSmoothed}
\alias{filterSmoothed}
\title{First, remove claims where linear and non-linear terms appear in the same claim}
\usage{
filterSmoothed(b, modelList)
}
\description{
First, remove claims where linear and non-linear terms appear in the same claim
}
\keyword{internal}
