% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.psem}
\alias{fitted.psem}
\title{Get fit values from a fitted piecewise SEM}
\usage{
\method{fitted}{psem}(object)
}
\arguments{
\item{object}{A  [psem] object}
}
\value{
A data.frame of fitted values
}
\description{
Get fit values from a fitted piecewise SEM
}
\examples{
data(keeley)
fit <- psem(
  lm(abiotic ~ distance, data=keeley),
  lm(rich ~ abiotic + hetero, data=keeley),
  lm(hetero ~ distance, data=keeley),
  data = keeley)

fit_vals <- fitted(fit)
          
head(fit_vals)
}
