% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{fixCatDir}
\alias{fixCatDir}
\title{Flip independence claims so categorical variables are not the response}
\usage{
fixCatDir(b, modelList)
}
\description{
Flip independence claims so categorical variables are not the response
}
\keyword{internal}
