% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get coefficients from linear regression}
\usage{
getCoefficients(model, data = NULL, test.type = "II")

getCoefficients(model, data = NULL, test.type = "II")
}
\description{
Get coefficients from linear regression

Get coefficients from linear regression
}
\keyword{internal}
