% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDAG.R
\name{getDAG}
\alias{getDAG}
\title{Generate adjacency matrix from list of structural equations}
\usage{
getDAG(modelList)
}
\arguments{
\item{modelList}{A list of structural equations}
}
\description{
Generate adjacency matrix from list of structural equations
}
