% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLchisq.R
\name{getSatModels}
\alias{getSatModels}
\title{Get saturated model by reinserting all excluded paths}
\usage{
getSatModels(b, modelList, data)
}
\description{
Get saturated model by reinserting all excluded paths
}
\keyword{internal}
