% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{GetSDx}
\alias{GetSDx}
\title{Get standard deviation of predictor variables}
\usage{
GetSDx(model, modelList, data, standardize = "scale")

GetSDx(model, modelList, data, standardize = "scale")
}
\description{
Get standard deviation of predictor variables

Get standard deviation of predictor variables
}
\keyword{internal}
