% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{GetSDy}
\alias{GetSDy}
\title{Properly scale standard deviations depending on the error distribution}
\usage{
GetSDy(model, data, standardize = "scale",
  standardize.type = "latent.linear")

GetSDy(model, data, standardize = "scale",
  standardize.type = "latent.linear")
}
\description{
Properly scale standard deviations depending on the error distribution

Properly scale standard deviations depending on the error distribution
}
\keyword{internal}
