% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSortedPsem}
\alias{getSortedPsem}
\title{Get a sorted psem object in DAG order}
\usage{
getSortedPsem(object)
}
\arguments{
\item{object}{A fit [psem] object}
}
\value{
A new [psem] object, without the data.
}
\description{
Takes a [psem] object, pulls out the
DAG, and then sorts the psem object into the order
of the DAG (from exogenous to terminal endogenous
variable) for use by other functions. Note: removes
correlated errors.
}
