% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{handleCategoricalCoefs}
\alias{handleCategoricalCoefs}
\title{Handles putting categorical variables into coefficient tables
for easy use in path analysis}
\usage{
handleCategoricalCoefs(
  ret,
  model,
  data,
  test.statistic = "F",
  test.type = "II"
)
}
\description{
Handles putting categorical variables into coefficient tables
for easy use in path analysis
}
\keyword{internal}
