% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infCrit.R
\name{infCrit}
\alias{infCrit}
\title{Information criterion values for SEM}
\usage{
infCrit(modelList, Cstat, add.claims = NULL, direction = NULL,
  conserve = FALSE, conditional = FALSE, .progressBar = FALSE)
}
\arguments{
\item{modelList}{a list of structural equations}

\item{...}{additional arguments}
}
\description{
Information criterion values for SEM
}
