% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{intCat}
\alias{intCat}
\title{Uses emmeans to get the mean at different levels of factors}
\usage{
intCat(model, catVarsInInt)
}
\description{
Uses emmeans to get the mean at different levels of factors
}
\keyword{internal}
