% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{intTrend}
\alias{intTrend}
\title{Uses emtrends to get the slope at different levels of factors}
\usage{
intTrend(model, catVarsInInt, contVarsInInt)
}
\description{
Uses emtrends to get the slope at different levels of factors
}
\keyword{internal}
