% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeley.R
\docType{data}
\name{keeley}
\alias{keeley}
\title{Data set from Keeley et al.}
\format{A \code{data.frame} with 90 observations of 8 variables.}
\usage{
keeley
}
\description{
\itemize{
\item{distance}{Distance to coast}
\item{elev}{Elevation from sea level}
\item{abiotic}{Abiotic favorability}
\item{age}{Age of stand before fire}
\item{hetero}{Plot heterogeneity}
\item{firesev}{Severity of fire}
\item{cover}{Cover of plants}
\item{rich}{Plant species richness}
}
}
\examples{
head(keeley)
}
\keyword{data}
