% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{KRp}
\alias{KRp}
\title{Recompute P-values using Kenward-Rogers approximation}
\usage{
KRp(model, vars, data, intercepts = FALSE)
}
\description{
Recompute P-values using Kenward-Rogers approximation
}
\keyword{internal}
