% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroup.R
\name{multigroup}
\alias{multigroup}
\title{Multigroup Analysis for Piecewise SEM}
\usage{
multigroup(modelList, data, group)
}
\arguments{
\item{group}{}
}
\value{

}
\description{
Multigroup Analysis for Piecewise SEM
}
