% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.psem}
\alias{predict.psem}
\title{Generate predictions from a fit piecewise SEM}
\usage{
\method{predict}{psem}(object, newdata = NULL, se.fit = FALSE, interval = "none", ...)
}
\arguments{
\item{object}{An object of class [psem]}

\item{newdata}{Optional. A data frame with new values
of exogenous variables}

\item{se.fit}{Do you want to incorporate coefficient 
standard errors? Currently defaults to FALSE as it has
not been implemented yet.}

\item{interval}{What type of interval do we want to 
report - fit, prediction, or none. Currently not 
implemented, so "none" is defaulted. Arugment here for
later development.}

\item{...}{Other arguments to `predict` functions used by
pieces of the `psem` object.}
}
\value{
A data.frame of predicted values
}
\description{
Generate predictions from a fit piecewise SEM
}
\examples{
data(keeley)
fit <- psem(
  lm(abiotic ~ distance, data=keeley),
  lm(rich ~ abiotic + hetero, data=keeley),
  lm(hetero ~ distance, data=keeley),
  data = keeley)

fit_vals <- predict(fit)

new_predictions <- 
  predict(fit, 
          newdata=data.frame(distance=c(30, 50)))
          
head(fit_vals)
head(new_predictions)
}
