% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_piecewiseSEM_newdata}
\alias{predict_piecewiseSEM_newdata}
\title{Generate fit values newdata with psem model}
\usage{
predict_piecewiseSEM_newdata(modList, newdata, lhs_sorted, ...)
}
\description{
Generate fit values newdata with psem model
}
\keyword{internal}
