% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_piecewiseSEM_nodata}
\alias{predict_piecewiseSEM_nodata}
\title{Extract fit values from models making up psem}
\usage{
predict_piecewiseSEM_nodata(modList, ...)
}
\description{
Extract fit values from models making up psem
}
\keyword{internal}
