% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_psem.R
\name{print.anova.psem}
\alias{print.anova.psem}
\title{Print anova}
\usage{
\method{print}{anova.psem}(x, ...)
}
\arguments{
\item{x}{an object of class anova.psem}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print anova
}
