% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{print.basisSet}
\alias{print.basisSet}
\title{Print basis set}
\usage{
\method{print}{basisSet}(x, ...)
}
\arguments{
\item{x}{a basis set}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print basis set
}
