% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroup.R
\name{print.multigroup.psem}
\alias{print.multigroup.psem}
\title{Print multigroup}
\usage{
\method{print}{multigroup.psem}(x, ...)
}
\arguments{
\item{x}{an object to print}

\item{...}{additional arguments to print}
}
\description{
Print multigroup
}
