% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{print.psem}
\alias{print.psem}
\title{Print psem}
\usage{
\method{print}{psem}(x, ...)
}
\arguments{
\item{x}{an object of class psem}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print psem
}
