% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.psem}
\alias{print.summary.psem}
\title{Print summary}
\usage{
\method{print}{summary.psem}(x, ...)
}
\arguments{
\item{x}{an object of class summary.psem}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print summary
}
