% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{removeCerror}
\alias{removeCerror}
\title{Remove correlated errors from the basis set}
\usage{
removeCerror(b, formulaList)
}
\description{
Remove correlated errors from the basis set
}
\keyword{internal}
