% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeData}
\alias{removeData}
\title{Remove data from the model list}
\usage{
removeData(modelList, formulas = 0)
}
\description{
formulas = 0, keep everything
formulas = 1, remove all formulas including correlated errors
formulas = 2, remove only formula but keep correlated errors
formulas = 3, remove correlated errors but keep formula
}
\keyword{internal}
