% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialResid.R
\name{resid.lme}
\alias{resid.lme}
\title{Get residuals from innermost grouping of mixed models (replicate-level)}
\usage{
\method{resid}{lme}(model)
}
\description{
Get residuals from innermost grouping of mixed models (replicate-level)
}
\keyword{internal}
