% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.psem}
\alias{residuals.psem}
\title{Residual values from fit models}
\usage{
\method{residuals}{psem}(object, ...)
}
\arguments{
\item{object}{a \code{\link{psem}} object}
}
\description{
Residual values from fit models
}
