% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{reverseNonLin}
\alias{reverseNonLin}
\title{If intermediate endogenous variables are nonlinear, return both directions}
\usage{
reverseNonLin(b, modelList, amat)
}
\description{
If intermediate endogenous variables are nonlinear, return both directions
}
\keyword{internal}
