% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared.merMod}
\alias{rsquared.merMod}
\title{R^2 for phylolm objects}
\usage{
rsquared.merMod(model)
}
\description{
R^2 for phylolm objects
}
\keyword{R^2}
\keyword{for}
\keyword{internal}
\keyword{merMod}
\keyword{objects}
