% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{safe_predict}
\alias{safe_predict}
\title{Make sure predict arguments are passed as needed}
\usage{
safe_predict(mod, ...)
}
\description{
Make sure predict arguments are passed as needed
}
\keyword{internal}
