% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{scaleFam}
\alias{scaleFam}
\title{Properly scale standard deviations depending on the error distribution}
\usage{
scaleFam(y, model, newdata, standardize = "scale",
  standardize.type = "latent.linear")
}
\description{
Properly scale standard deviations depending on the error distribution
}
