% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{scaleGLM}
\alias{scaleGLM}
\title{Compute standard deviation or relevant range of response for GLMs}
\usage{
scaleGLM(model, standardize = "scale", standardize.type = "Menard.OE")
}
\description{
Compute standard deviation or relevant range of response for GLMs
}
