% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{scaleInt}
\alias{scaleInt}
\title{Calculate standard deviation or relevant range for interaction terms}
\usage{
scaleInt(model, newdata, standardize)

scaleInt(model, newdata, standardize)
}
\description{
Calculate standard deviation or relevant range for interaction terms

Calculate standard deviation or relevant range for interaction terms
}
\keyword{internal}
