% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{sdFam}
\alias{sdFam}
\title{Properly scale standard deviations depending on the error distribution}
\usage{
sdFam(x, model, newdata)
}
\description{
Properly scale standard deviations depending on the error distribution
}

