% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{sem.coefs}
\alias{sem.coefs}
\title{sem.coefs}
\usage{
sem.coefs(modelList, data = NULL, standardize = "none",
  corr.errors = NULL, intercept = FALSE)
}
\arguments{
\item{modelList}{a \code{list} of regressions representing the structural equation model}

\item{data}{a \code{data.frame} used to construct the structured equations}

\item{standardize}{whether coefficients should be scaled by their standard deviations ("scale") or by
their ranges ("range")}

\item{corr.errors}{a vector of variables with correlated errors (separated by "~~")}

\item{intercept}{whether intercepts should also be returned. Default is \code{FALSE}}
}
\description{
Standardized coefficients for linear models (old)
}
