% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{sem.fisher.c}
\alias{sem.fisher.c}
\title{sem.fisher.c}
\usage{
sem.fisher.c(modelList, data, corr.errors = NULL, add.vars = NULL,
  grouping.vars = NULL, grouping.fun = mean, adjust.p = FALSE,
  basis.set = NULL, pvalues.df = NULL, model.control = NULL,
  .progressBar = TRUE)
}
\description{
sem.fisher.c
}
\keyword{internal}
