% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{sem.fit}
\alias{sem.fit}
\title{sem.fit}
\usage{
sem.fit(modelList, data, conditional = FALSE, corr.errors = NULL,
  add.vars = NULL, grouping.vars = NULL, grouping.fun = mean,
  adjust.p = FALSE, basis.set = NULL, pvalues.df = NULL,
  model.control = NULL, .progressBar = TRUE)
}
\description{
sem.fit
}
\keyword{internal}
