% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{sem.model.fits}
\alias{sem.model.fits}
\title{sem.model.fits}
\usage{
sem.model.fits(...)
}
\arguments{
\item{...}{a linear model or list of models}
}
\description{
Goodness-of-fit statistics for linear models (old)
}
