% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shipley}
\alias{shipley}
\title{Data set from Shipley (2006)}
\format{
A \code{data.frame} with 1900 observations of 9 variables.
\describe{
\item{site}{Site of observation}
\item{tree}{Individual tree of observation}
\item{lat}{Latitude}
\item{year}{Year of observation}
\item{Date}{Julian date of first bud burst}
\item{DD}{Cumulative degree days until first bud burst}
\item{Growth}{Increase in stem diameter}
\item{Survival}{Proportional survival}
\item{Live}{Alive (1) or dead (0)}
}
}
\usage{
shipley
}
\description{
Data set from Shipley (2006)
}
\keyword{data}
