\docType{data}

\name{shipley2009}

\alias{shipley2009}

\title{
Shipley (2009) Tree Data
}

\description{
Hypothetical data from Shipley (2009).
}

\details{
Simulated data from a hypothetical study beginning in 1970 in which 20 sites \code{site} are chosen differing in latitude \code{lat}. Five individual trees \code{tree} of a particular species are chosen within each site. Each tree is followed every second year \code{year} until 2006 or until it dies (thus, repeated measures). At each site, in each sample year, and for each living individual you measure the cumulative degree days \code{DD} until bud break, the Julian date (day of year) \code{Date} of bud break, the increase in stem diameter per tree \code{Growth}, and a binary variable indicating survival (1) or death (0) \code{Survival} during the subsequent growing season
}

\references{
Shipley, Bill. "Confirmatory path analysis in a generalized multilevel context." Ecology 90.2 (2009): 363-368.

http://www.esapubs.org/archive/ecol/E090/028/Shipley.dat
}

\author{
Jon Lefcheck
}

\keyword{ data }

\keyword{ datasets }