\docType{data}
\name{shipley2013}
\alias{shipley2013}
\title{
Shipley (2013) Simulated Data
}
\description{
Hypothetical data from Shipley (2013).
}
\details{
Simulated data containing 250 observations involving 10 individuals in each of 25 \code{species}. Five variables (\code{x1},\code{x2}, etc.) have been measured on each individual.
}
\references{
Shipley, Bill. "Confirmatory path analysis in a generalized multilevel context." Ecology 90.2 (2009): 363-368.
}
\author{
Jon Lefcheck
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ data }
\keyword{ datasets }