% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{specifyDir}
\alias{specifyDir}
\title{Remove items from the basis set whose direction is a priori specified}
\usage{
specifyDir(b, direction)
}
\description{
Remove items from the basis set whose direction is a priori specified
}
\keyword{internal}
