% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{unstdCoefs}
\alias{unstdCoefs}
\title{Get raw (undstandardized) coefficients from model}
\usage{
unstdCoefs(modelList, data = NULL, test.type = "III",
  intercepts = FALSE)

unstdCoefs(modelList, data = NULL, test.type = "III",
  intercepts = FALSE)
}
\description{
Get raw (undstandardized) coefficients from model

Get raw (undstandardized) coefficients from model
}
\keyword{internal}
