% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{update.psem}
\alias{update.psem}
\title{Update psem model object with additional values.}
\usage{
\method{update}{psem}(object, ...)
}
\arguments{
\item{object}{a psem object}

\item{...}{additional arguments to update}
}
\description{
Update psem model object with additional values.
}
\examples{
mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley
)

update(mod, firesev ~ age + cover)

}
