% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.psem}
\alias{anova.psem}
\title{Compare models using Fisher's C scores}
\usage{
\method{anova}{psem}(object, object2 = NULL, fun = Anova, ...)
}
\arguments{
\item{object}{a \code{\link{psem}} object}

\item{object2}{an optional second \code{\link{psem}} object}

\item{fun}{what anova function to use. Default is car::Anova}

\item{...}{options for anova functions}
}
\description{
Compare models using Fisher's C scores
}
