% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_psem.R
\name{anova.psem}
\alias{anova.psem}
\title{ANOVA and chi-squared difference test for model comparison}
\usage{
\method{anova}{psem}(mod, mod2 = NULL, digits = 3,
  anovafun = car::Anova, ...)
}
\arguments{
\item{mod}{a \code{psem} object}

\item{mod2}{a \code{psem} object for comparison. Defaults to NULL to allow for a LRT or F tables of all fit model pieces.}

\item{digits}{number of digits to round results. Default is 3}

\item{anovafun}{The function used for ANOVA. Defaults to \code{\link{car::Anova}}}

\item{...}{additional arguments passed to \code{anovafun}}
}
\value{
an F, LRT, or other table for a single model, or a list of
 comparisons between multiple models
}
\description{
Compute analysis of variance table for one or more structural equation models.
}
\examples{
mod1 <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley
)

# get type II Anova
anova(mod1)

# conduct LRT
mod2 <- psem(
  lm(rich ~ cover, data = keeley),
  lm(cover ~ firesev, data = keeley),
  age ~ 1,
  data = keeley
)

anova(mod1, mod2)

}
\seealso{
The model fitting function \code{\link{psem}}
}
\author{
Jon Lefcheck <LefcheckJ@si.edu>, Jarrett Byrnes <jarrett.byrnes@umb.edu>
}
