% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{Dag}
\alias{Dag}
\title{Generate adjacency matrix from list of structural equations}
\usage{
Dag(formulaList)
}
\arguments{
\item{formulaList}{a list of formulae corresponding to structural equations}
}
\description{
Generate adjacency matrix from list of structural equations
}
