% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dataTrans}
\alias{dataTrans}
\title{Transform variables based on model formula and store in new data frame}
\usage{
dataTrans(formula., data)
}
\description{
Transform variables based on model formula and store in new data frame
}
\keyword{internal}
