% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\name{evaluateClasses}
\alias{evaluateClasses}
\title{Evaluate model classes and stop if unsupported model class}
\usage{
evaluateClasses(x)
}
\arguments{
\item{x}{a list of structural equations or a model object}
}
\description{
Evaluate model classes and stop if unsupported model class
}
