% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{filter.exogenous}
\alias{filter.exogenous}
\title{filter.exogenous}
\usage{
filter.exogenous(modelList, basis.set, corr.errors = NULL, add.vars = NULL)
}
\description{
filter.exogenous
}
\keyword{internal}
