% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherC.R
\name{fisherC}
\alias{fisherC}
\title{Summarize tests of directed separation using Fisher's C statistic}
\usage{
fisherC(dTable, add.claims = NULL, direction = NULL, conserve = FALSE,
  conditional = FALSE, .progressBar = FALSE)
}
\arguments{
\item{dTable}{a data.frame of evaluated independence claims}
}
\description{
Summarize tests of directed separation using Fisher's C statistic
}

