% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get coefficients from linear regression}
\usage{
getCoefficients(model, data = NULL, test.type = "III")
}
\description{
Get coefficients from linear regression
}
\keyword{internal}
