% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialResid.R
\name{getResidModels}
\alias{getResidModels}
\title{Identify models with correlated errors and return modified versions}
\usage{
getResidModels(vars, modelList, data)
}
\description{
Identify models with correlated errors and return modified versions
}
\keyword{internal}
