% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{GetSDx}
\alias{GetSDx}
\title{Get standard deviation of predictor variables}
\usage{
GetSDx(model, modelList, data, standardize = "scale")
}
\description{
Get standard deviation of predictor variables
}
\keyword{internal}
