% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keeley}
\alias{keeley}
\title{Data set from Grace & Keeley (2006)}
\format{
A \code{data.frame} with 90 observations of 8 variables.
\describe{
\item{distance}{Distance to coast}
\item{elev}{Elevation from sea level}
\item{abiotic}{Abiotic favorability}
\item{age}{Age of stand before fire}
\item{hetero}{Plot heterogeneity}
\item{firesev}{Severity of fire}
\item{cover}{Cover of plants}
\item{rich}{Plant species richness}
}
}
\usage{
keeley
}
\description{
Data set from Grace & Keeley (2006)
}
\keyword{data}
