% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listFormula}
\alias{listFormula}
\title{Recompute P-values using Kenward-Rogers approximation}
\usage{
listFormula(modelList, formulas = 0)
}
\description{
Recompute P-values using Kenward-Rogers approximation
}
\keyword{Get}
\keyword{If}
\keyword{`sem`}
\keyword{a}
\keyword{formula}
\keyword{from}
\keyword{internal}
\keyword{is}
\keyword{list}
\keyword{non-evaluated}
\keyword{object}
\keyword{of}
\keyword{out}
\keyword{remove}
\keyword{take}
\keyword{true,}
