% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroup.R
\name{multigroup}
\alias{multigroup}
\title{Multigroup Analysis for Piecewise SEM}
\usage{
multigroup(modelList, group)
}
\arguments{
\item{modelList}{a list of structural equations}

\item{group}{the name of the grouping variable in quotes}
}
\description{
Multigroup Analysis for Piecewise SEM
}
\examples{
data(meadows)

jutila <- psem(
lm(rich ~ elev + mass, data = meadows),
lm(mass ~ elev, data = meadows)
)

jutila.multigroup <- multigroup(jutila, group = "grazed")

jutila.multigroup

}
\author{
Jon Lefcheck <LefcheckJ@si.edu>
}
