% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisSet.R
\name{reverseAddVars}
\alias{reverseAddVars}
\title{Replace transformations in the basis set by cycling through neighbors and applying
transformations in order of how variables are treated in the child nearest to current node}
\usage{
reverseAddVars(b, modelList, amat)
}
\description{
Replace transformations in the basis set by cycling through neighbors and applying
transformations in order of how variables are treated in the child nearest to current node
}
\keyword{Reverse}
\keyword{added}
\keyword{internal}
\keyword{variables}
