% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R, R/coefs_old_feb2019.R
\name{stdCoefs}
\alias{stdCoefs}
\title{Calculate standardized regression coefficients}
\usage{
stdCoefs(modelList, data = NULL, standardize = "scale",
  standardize.type = "latent.linear", test.type = "II",
  intercepts = FALSE)

stdCoefs(modelList, data = NULL, standardize = "scale",
  standardize.type = "latent.linear", test.type = "II",
  intercepts = FALSE)
}
\description{
Calculate standardized regression coefficients

Calculate standardized regression coefficients
}
\keyword{internal}
