% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{stdCoefs}
\alias{stdCoefs}
\title{Calculate standardized regression coefficients}
\usage{
stdCoefs(
  modelList,
  data = NULL,
  standardize = "scale",
  standardize.type = "latent.linear",
  test.statistic = "F",
  test.type = "II",
  intercepts = FALSE
)
}
\description{
Calculate standardized regression coefficients
}
\keyword{internal}
