% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSep.R
\name{testBasisSetElements}
\alias{testBasisSetElements}
\title{Evaluate conditional independence claim from the basis set}
\usage{
testBasisSetElements(i, b, modelList, data, conditioning, .progressBar, pb)
}
\description{
Evaluate conditional independence claim from the basis set
}
\keyword{internal}
