% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{unstdCoefs}
\alias{unstdCoefs}
\title{Get raw (understandardized) coefficients from model}
\usage{
unstdCoefs(modelList, data = NULL, intercepts = FALSE)
}
\description{
Get raw (understandardized) coefficients from model
}

