% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{unstdCoefs}
\alias{unstdCoefs}
\title{Get raw (undstandardized) coefficients from model}
\usage{
unstdCoefs(
  modelList,
  data = NULL,
  test.statistic = "F",
  test.type = "II",
  intercepts = FALSE
)
}
\description{
Get raw (undstandardized) coefficients from model
}
\keyword{internal}
